moved code that populates our profile into its own file and made it into a function
